+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2025-03-09
+
+### Added
+
+- Add configuration update method to update the configuration when the settings are changed
+
+### Changed
+
+- Improve localization strings for the extension
+
 ## [1.1.0] - 2025-03-09
 
 ### Added
@@ -21,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of Dart Barrel File extension
 
-[Unreleased]: https://github.com/ManuelGil/ovsx-dart-barrel/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/ManuelGil/ovsx-dart-barrel/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/ManuelGil/ovsx-dart-barrel/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/ManuelGil/ovsx-dart-barrel/compare/v1.0.0...v1.1.0
 [1.0.0]: https://githttps://github.com/ManuelGil/ovsx-dart-barrel/releases/tag/v1.0.0